requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==5.5.3
google-generativeai==0.5.4
openai==1.12.0
qdrant-client==1.8.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
python-dotenv==1.0.0
langchain-community==0.0.19
pydantic==2.5.0
mangum==0.17.0
httpx==0.25.2
